Finish dropdown setup
