+++ index.html
@@ -40,35 +40,35 @@
           
           <div class="input">
             <label for="route-select">Your route</label><br>            
-            <select name="route" id="route-select">
+            <select name="route" id="route">
               <option value="">Please choose your route</option>
             </select>
           </div>
           
           <div class="input">
             <label for="direction-select">Trip direction</label><br>            
-            <select name="direction" id="direction-select" disabled>
+            <select name="direction" id="direction" disabled>
               <option value="">Please choose your route first</option>
             </select>
           </div>
           
           <div class="input">
             <label for="weekday-select">Day of week</label><br>            
-            <select name="weekday" id="weekday-select" disabled>
+            <select name="weekday" id="weekday" disabled>
               <option value="">Please choose the preceding options first</option>
             </select>
           </div>
           
           <div class="input">
             <label for="time-select">Time of day</label><br>            
-            <select name="time" id="time-select" disabled>
+            <select name="time" id="time" disabled>
               <option value="">Please choose the preceding options first</option>
             </select>
           </div>
           
           <div class="input">
             <label for="stop-select">Your stop</label><br>            
-            <select name="stop" id="stop-select" disabled>
+            <select name="stop" id="stop" disabled>
               <option value="">Please choose the preceding options first</option>
             </select>
           </div>
