+++ js/main.js
@@ -1,4 +1,5 @@
 import * as utils from './utils.js';
+import * as dropdown from './dropdown.js';
 import { initMap } from './map.js';
 
 // Events in event bus:
@@ -15,25 +16,16 @@ const arrivalsAndStops = await utils.getGeojsonCollection('arrivals_and_stops.js
 initMap(mapElement, septaNetwork);
 
 // Populate route dropdown and set up listener
-utils.populateDropdownOptions('route-select', arrivalsAndStops, 'route_id', 'Please choose your route');
-utils.setupDropdownChangeListener('route-select', 'routeSelected', events);
+dropdown.populateDropdown(events, arrivalsAndStops, 'route', 'route_id', 'Please choose your route');
 
 // Populate trip direction dropdown based on selected route
-utils.cascadeDropdown(events, arrivalsAndStops, 'routeSelected', 'route_id', 'trip_headsign', 'direction-select', 'Please choose your trip direction');
-utils.setupDropdownChangeListener('direction-select', 'directionSelected', events);
+dropdown.cascadeDropdown(events, 'routeData', 'direction', 'trip_headsign', 'Please choose your trip direction');
+
+// Populate day of week dropdown based on previous selections
+dropdown.cascadeDropdown(events, 'directionData', 'weekday', 'week_period', 'Please choose the day of the week');
     
 // Populate day of week dropdown based on previous selections
-utils.cascadeDropdown(events, arrivalsAndStops, 'directionSelected', 'trip_headsign', 'week_period', 'weekday-select', 'Please choose your day of the week');
-utils.setupDropdownChangeListener('weekday-select', 'weekdaySelected', events);
-
-// Populate day of week dropdown based on previous selections
-utils.cascadeDropdown(events, arrivalsAndStops, 'weekdaySelected', 'week_period', 'time_of_day', 'time-select', 'Please choose your time of day');
-utils.setupDropdownChangeListener('time-select', 'timeSelected', events);
-
-// Populate day of week dropdown based on previous selections
-utils.cascadeDropdown(events, arrivalsAndStops, 'weekdaySelected', 'week_period', 'time_of_day', 'time-select', 'Please choose your time of day');
-utils.setupDropdownChangeListener('time-select', 'timeSelected', events);
+dropdown.cascadeDropdown(events, 'weekdayData', 'time', 'time_of_day', 'Please choose your stop');
 
 // Populate stop dropdown based on previous selections
-utils.cascadeDropdown(events, arrivalsAndStops, 'timeSelected', 'time_of_day', 'stop_name', 'stop-select', 'Please choose your stop');
-utils.setupDropdownChangeListener('stop-select', 'stopSelected', events);
+dropdown.cascadeDropdown(events, 'timeData', 'stop', 'stop_name', 'Please choose your stop');
